Share pane-target resolution with top-level kill
Top-level maw kill still used the vendored session resolver, so orphan panes addressed by title or worktree alias missed the #1502 tmux fallback. Move the pure pane-target resolver into shared code and let the kill plugin use it when no session matches, while preserving --pane index validation.

Constraint: maw kill <session> and maw kill <session> --pane N semantics must remain stable.

Rejected: route all top-level kill calls through cmdTmuxKill | that would change bare session kills into pane kills.

Confidence: high

Scope-risk: narrow

Directive: Keep pane title/worktree alias resolution shared between tmux kill and top-level kill; do not duplicate divergent fuzzy matching.

Tested: bun test test/isolated/kill-pane-range.test.ts test/isolated/tmux-kill-pane-target.test.ts; bun run build

Not-tested: live orphan-pane kill on Nat's mawjs session; avoid killing panes from this Codex lane.
